[功能][CSROP-4008] ToolShowcase 模組
